+++ python/lsst/ts/wep/task/cutOutDonutsBase.py
@@ -452,6 +452,70 @@ reducing the amount of donut mask dilation to {self.bkgDilationIter}"
         shiftFailureIdx = np.where(totalShifts > self.maxRecenterDistance)[0]
         return shiftFailureIdx
 
+    def addVisitLevelMetadata(self, exposure, inputDonutStamps, donutCatalog, defocalType):
+        """
+        Get visit level metadata and save in the donutStamps object.
+
+        Parameters
+        ----------
+        exposure : lsst.afw.image.Exposure
+            Exposure to get metadata from.
+        inputDonutStamps : DonutStamps
+            DonutStamps object to save metadata in.
+        donutCatalog : astropy.table.QTable
+            Source catalog for the pointing.
+        defocalType : enum 'DefocalType'
+            Defocal type of the donut image.
+
+        Returns
+        -------
+        DonutStamps
+            DonutStamps object with metadata added.
+        """
+        if inputDonutStamps.metadata is None:
+            inputDonutStamps.metadata = PropertyList()
+
+        visitInfo = exposure.getInfo().getVisitInfo()
+        detector = exposure.getDetector()
+        detectorName = detector.getName()
+        bandLabel = exposure.filter.bandLabel
+        cameraName = donutCatalog.meta['visit_info']['instrument_label']
+
+        instrument = getTaskInstrument(
+            cameraName,
+            detectorName,
+            self.instConfigFile,
+        )
+        inputDonutStamps.metadata["BANDPASS"] = bandLabel
+        inputDonutStamps.metadata["DFC_DIST"] = instrument.defocalOffset * 1e3
+        inputDonutStamps.metadata["DFC_TYPE"] = defocalType.value
+        inputDonutStamps.metadata["DET_NAME"] = detectorName
+        inputDonutStamps.metadata["VISIT"] = donutCatalog.meta['visit_info']['visit_id']
+        inputDonutStamps.metadata["MJD"] =  visitInfo.date.toAstropy().tai.mjd
+        inputDonutStamps.metadata["CAM_NAME"] = cameraName
+
+        # Save visit info
+        inputDonutStamps.metadata["BORESIGHT_ROT_ANGLE_RAD"] = (
+            donutCatalog.meta["visit_info"]["boresight_rot_angle"].to(u.rad).value
+        )
+        inputDonutStamps.metadata["BORESIGHT_PAR_ANGLE_RAD"] = (
+            donutCatalog.meta["visit_info"]["boresight_par_angle"].to(u.rad).value
+        )
+        inputDonutStamps.metadata["BORESIGHT_ALT_RAD"] = (
+            donutCatalog.meta["visit_info"]["boresight_alt"].to(u.rad).value
+        )
+        inputDonutStamps.metadata["BORESIGHT_AZ_RAD"] = (
+            donutCatalog.meta["visit_info"]["boresight_az"].to(u.rad).value
+        )
+        inputDonutStamps.metadata["BORESIGHT_RA_RAD"] = (
+            donutCatalog.meta["visit_info"]["boresight_ra"].to(u.rad).value
+        )
+        inputDonutStamps.metadata["BORESIGHT_DEC_RAD"] = (
+            donutCatalog.meta["visit_info"]["boresight_dec"].to(u.rad).value
+        )
+
+        return inputDonutStamps
+
     def cutOutStamps(self, exposure, donutCatalog, defocalType, cameraName):
         """
         Cut out postage stamps for sources in catalog.
@@ -477,7 +541,6 @@ reducing the amount of donut mask dilation to {self.bkgDilationIter}"
         detector = exposure.getDetector()
         detectorName = detector.getName()
         bandLabel = exposure.filter.bandLabel
-        visitId = exposure.getInfo().getVisitInfo().id
         catalogMeta = donutCatalog.meta
 
         # Run background subtraction
@@ -675,9 +738,8 @@ reducing the amount of donut mask dilation to {self.bkgDilationIter}"
             finalStamps.append(donutStamp)
 
         # Add additional information into metadata
-        catalogLength = len(donutCatalog)
         stampsMetadata = PropertyList()
-        stampsMetadata["VISIT"] = np.array([visitId] * catalogLength, dtype=int)
+
         # Save the donut flux as magnitude
         fluxLabel = next(
             (
@@ -702,26 +764,6 @@ reducing the amount of donut mask dilation to {self.bkgDilationIter}"
         stampsMetadata["CENT_DR"] = np.sqrt(
             donutCatalog["xShift"] ** 2 + donutCatalog["yShift"] ** 2
         )
-        # Save visit info
-        stampsMetadata["BORESIGHT_ROT_ANGLE_RAD"] = (
-            donutCatalog.meta["visit_info"]["boresight_rot_angle"].to(u.rad).value
-        )
-        stampsMetadata["BORESIGHT_PAR_ANGLE_RAD"] = (
-            donutCatalog.meta["visit_info"]["boresight_par_angle"].to(u.rad).value
-        )
-        stampsMetadata["BORESIGHT_ALT_RAD"] = (
-            donutCatalog.meta["visit_info"]["boresight_alt"].to(u.rad).value
-        )
-        stampsMetadata["BORESIGHT_AZ_RAD"] = (
-            donutCatalog.meta["visit_info"]["boresight_az"].to(u.rad).value
-        )
-        stampsMetadata["BORESIGHT_RA_RAD"] = (
-            donutCatalog.meta["visit_info"]["boresight_ra"].to(u.rad).value
-        )
-        stampsMetadata["BORESIGHT_DEC_RAD"] = (
-            donutCatalog.meta["visit_info"]["boresight_dec"].to(u.rad).value
-        )
-        stampsMetadata["MJD"] = donutCatalog.meta["visit_info"]["mjd"]
 
         if len(finalStamps) > 0:
             self.metadata[f"recenterFlags{defocalType.value.capitalize()}"] = list(
@@ -784,5 +826,12 @@ reducing the amount of donut mask dilation to {self.bkgDilationIter}"
         # Refresh to pull original metadata into stamps
         # Necessary when running full pipeline interactively.
         finalDonutStamps._refresh_metadata()
+
+        finalDonutStamps = self.addVisitLevelMetadata(
+            exposure,
+            finalDonutStamps,
+            donutCatalog,
+            defocalType,
+        )
 
         return finalDonutStamps