+++ tests/task/test_calcZernikesUnpairedTask.py
@@ -203,10 +203,13 @@ class TestCalcZernikeUnpaired(lsst.utils.tests.TestCase):
                 self.assertIn("cam_name", structNormal.zernikes.meta)
                 for k in ["intra", "extra"]:
                     dict_ = structNormal.zernikes.meta[k]
-                    self.assertIn("det_name", dict_)
-                    self.assertIn("visit", dict_)
-                    self.assertIn("dfc_dist", dict_)
-                    self.assertIn("band", dict_)
+                    if k == stamps.metadata['DFC_TYPE']:
+                        self.assertIn("det_name", dict_)
+                        self.assertIn("visit", dict_)
+                        self.assertIn("dfc_dist", dict_)
+                        self.assertIn("band", dict_)
+                    else:
+                        self.assertEqual(dict_, {})
 
                 # Turn on the donut stamp selector
                 task.doDonutStampSelector = True
@@ -237,7 +240,6 @@ class TestCalcZernikeUnpaired(lsst.utils.tests.TestCase):
 
                 # test null run
                 structNull = task.run([])
-                print(structNull)
 
                 for struct in [structNormal, structNull]:
                     # test that in accordance with declared connections,